% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grids}
\alias{grids}
\title{Spatial grid dataset.}
\format{
\code{grids}: A data frame of explanatory variables with 68757 rows and 11 columns.
}
\usage{
grids
}
\description{
Spatial datasets of explanatory variables at a 500-m resolution.
}
\author{
Yongze Song \email{yongze.song@curtin.edu.au}
}
\keyword{grids}
\keyword{variables}
