% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs}
\alias{obs}
\title{Spatial datasets of trace elements.}
\format{
\code{obs}: A data frame of trace elements with 614 rows and 5 variables
}
\usage{
obs
}
\description{
Spatial datasets of trace elements.
}
\author{
Yongze Song \email{yongze.song@curtin.edu.au}
}
\keyword{dataset}
\keyword{elements}
\keyword{trace}
