% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvoutlier.R
\name{rmvoutlier}
\alias{rmvoutlier}
\title{Removing outliers.}
\usage{
rmvoutlier(x, coef = 2.5)
}
\arguments{
\item{x}{A vector of a variale}

\item{coef}{A number of the times of standard deviation. The default value is 2.5.}
}
\value{
Location of outliers in the vector
}
\description{
Function for removing outliers.
}
\examples{
data("obs")
obs$Cr_ppm <- log(obs$Cr_ppm)
krm <- rmvoutlier(obs$Cr_ppm)
krm

}
