% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_vars_fda}
\alias{sample_vars_fda}
\title{Spatial datasets of the first dimension variables of trace elements.}
\format{
\code{sample_vars_fda}: A list with 8 elements
}
\usage{
sample_vars_fda
}
\description{
Spatial datasets of the first dimension variables of trace elements.
}
\author{
Yongze Song \email{yongze.song@curtin.edu.au}
}
\keyword{dataset}
\keyword{elements}
\keyword{factors}
\keyword{trace}
