% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif.R
\name{vif}
\alias{vif}
\title{Fast calculation of the variance inflation factor (VIF)}
\usage{
vif(x)
}
\arguments{
\item{x}{A data.frame of explanatory variables}
}
\value{
Variance inflation factor (VIF) values of variables
}
\description{
Function for fast calculation of the variance inflation factor (VIF)
}
\examples{
data("sample_vars_sda")
x <- sample_vars_sda$Elevation[, sample(55, 10)]
vif(x)

}
